app/requirements.txt
anyio==3.6.2
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
cryptography==39.0.2
easycor==1.7.1
ecdsa==0.18.0
fastapi==0.95.0
h11==0.14.0
httptools==0.5.0
idna==3.4
passlib==1.7.4
psycopg2-binary==2.9.6
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.5
PyMySQL==1.0.3
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.7
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
Jinja2==3.1.4
redis==5.0.7
requests==2.32.3
aioredis==1.3.1 